+++ CHANGELOG.md
@@ -23,6 +23,9 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 * A new global flag `--no-integrate-operation` lets you run a command without
   impacting the repo state or the working copy.
 
+* A new config option `diff.git.show-path-prefix` can be used to suppress the
+  `a/` and `b/` path prefixes in the `diff --git` output.
+
 ## [0.40.0] - 2026-04-01
 
 ### Release highlights